backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
